+++ tests/Html/Builder/BuilderOptionsTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Yajra\DataTables\Html\Tests;
+namespace Yajra\DataTables\Html\Tests\Html\Builder;
 
 use PHPUnit\Framework\Attributes\Test;
 use Yajra\DataTables\Html\Builder;
 use Yajra\DataTables\Html\Column;
+use Yajra\DataTables\Html\Tests\TestCase;
 
 class BuilderOptionsTest extends TestCase
 {
