+++ tests/Html/Builder/BuilderTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yajra\DataTables\Html\Tests;
+namespace Yajra\DataTables\Html\Tests\Html\Builder;
 
 use Illuminate\Support\HtmlString;
 use PHPUnit\Framework\Attributes\Test;
@@ -9,6 +9,7 @@ use Yajra\DataTables\Html\Column;
 use Yajra\DataTables\Html\ColumnDefinition;
 use Yajra\DataTables\Html\ColumnDefinitions;
 use Yajra\DataTables\Html\Editor\Editor;
+use Yajra\DataTables\Html\Tests\TestCase;
 
 class BuilderTest extends TestCase
 {
