+++ tests/Html/Column/ColumnDefinitionTest.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Yajra\DataTables\Html\Tests;
+namespace Yajra\DataTables\Html\Tests\Html\Column;
 
 use PHPUnit\Framework\Attributes\Test;
 use Yajra\DataTables\Html\ColumnDefinition;
+use Yajra\DataTables\Html\Tests\TestCase;
 
 class ColumnDefinitionTest extends TestCase
 {
