+++ tests/Html/Column/ColumnTest.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Yajra\DataTables\Html\Tests;
+namespace Yajra\DataTables\Html\Tests\Html\Column;
 
 use PHPUnit\Framework\Attributes\Test;
 use Yajra\DataTables\Html\Column;
+use Yajra\DataTables\Html\Tests\TestCase;
 
 class ColumnTest extends TestCase
 {
