+++ tests/Html/Editor/EditorTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Yajra\DataTables\Html\Tests;
+namespace Yajra\DataTables\Html\Tests\Html\Editor;
 
 use PHPUnit\Framework\Attributes\Test;
 use Yajra\DataTables\Html\Editor\Editor;
 use Yajra\DataTables\Html\Editor\Fields\Text;
+use Yajra\DataTables\Html\Tests\TestCase;
 
 class EditorTest extends TestCase
 {
